Quickstart: Revert back to "python3 -m pip" invocation
I intentionally did not use a "pip3" invocation as it is not as
portable.  This is in-line with the official Python documentation, which
also uses the "python3 -m pip" invocation to avoid ambiguity:

  https://docs.python.org/3/installing/index.html

For example, some Python 3-only environments, like those produced with
Conda, don't have a pip3 command; they just have pip since there's only
one Python version available.