Quickstart: Rephrase summary of the build to include the word "build"
I think it's important to distinguish between "augur", which is a
software toolkit, and "the build", which is code + process that uses
that toolkit.  Conceptually we're not just "running augur" when running
a build.  Augur on it's own is just a collection of tools; the build is
what assembles those tools into a meaningful thing.