Add an introduction to augur page
Briefly describes augur as the basis for Nextstrain's bioinformatics and
its philosophy of composition.

It's possible we should provide a short, one-sentence description of
each command listed on this page, but I left that out for now in the
interest of having a document sooner than later.  Each command will be
described in more detail in the (linked) Commands page still to come.

This standardizes on "augur command" instead of "augur module" to more
closely match the language of the command-line.  Existing doc pages
which referred to the augur introduction or augur modules are updated.