Quickstart: Switch to using the zika-tutorial build
This is a simpler repo, put together by John, that starts with what we
imagine most people will have: a FASTA file of sequences and a CSV file
of metadata.  Using the tutorial repo in the quickstart means that it
matches the full zika tutorial, which is a nice progression.