Tutorials: Standardize on the noun "build" instead of mixing it with "site"
Everywhere else we've pretty uniformly used the noun "build" to refer to
the whole set of things and tasks that are involved with bringing a
pathogen into Nextstrain.  The tutorial documents introduced a new noun,
"site", to seemingly refer to the same thing.  I think this is
confusing, especially since "build" (the noun) was still used as well.

A "site" would seem to me to be some alternative to nextstrain.org, or
maybe a "community site", i.e. nextstrain.org/community/organization/….
These tutorials don't aim to cover either of those topics.