Automatically determine the last modification date of Markdown files from git
This frees us from having to remember to bump the date in the
frontmatter when we make changes.

Only documents with no "date" field in the frontmatter will use the date
from git.  This allows some documents to set a fixed explicit date, such as
"February 2018", if they so desire.  Blog posts and reports, for
example, may want the timestamp to ignore minor updates made after their
initial posting.