Distinguish post dates from last modified dates in the page heading
Some pages used the date to indicate when the initial posting was made,
others used it to indicate when a page was last updated.  Since the
latter usage is now done automatically, we can infer that explicit dates
are only "post dates" and omit the indication that they're a "Last
modified" timestamp.