Splash page: Describe Nextstrain as a bioinformatics and data viz toolkit too
Nextstrain is both the real-time analyses displayed on nextstrain.org
*and* the software tools we use to perform and present those analyses.
Recently we've come around to acknowledging that duality and describing
both sets of things in more places.  The splash page is another great
place to do that.

This grew out of conversation on #nextstrain-general starting with
<https://bedfordlab.slack.com/archives/C7SDVPBLZ/p1535654872000100> and
includes edits made by Trevor.