configure to source the static site from this repo
The static site is now located in this repo
(previously it was seperate, and the built output,
i.e. "/public" was uploaded to a S3 bucket).

This repo now builds the static site (via set-up.sh)
rather than fetching from S3. It is available in
"./static-site/public" (so the server paths were
slightly changed).

Additionally, the scripts in the static site relating
to heroku, S3 etc are removed.

Note that building the static site requires py27, so
the Heroku & Travis configs are modified.