remove lastCommitDate
Doc pages / blog posts render the "date modified" in the header. This normally comes from the "date" field in the markdown YAML. If this doesn't exist we used to get the "lastCommitDate" of the file. This fails on Heroku as it's not using git repositories. We now don't display a date if the date field is not specified in the markdown.