remove chalk library from auspice bundle
Previously, due to `utils.js` being in `./auspice/client` the auspice extension machinery would transform this file (and the libraries it imported) and include it in `bundle.js`. This wasn't noticed on Chrome / Firefox / Safari (the bundled code isn't actually called) but caused a parsing error on IE11 due to the `chalk` library using `for (const x of ...` statements which aren't allowed in IE11.

This commit restores functionality on IE11 and also reduces the bundle size of extended versions of auspice.