Order datasets selectors
Previously the dataset selectors were not in the desired order due to the default value appearing first. This commit fixes this for the nextstrain.org server. Note that this functionality will be moved to the client when the new server API is implemented (see https://github.com/nextstrain/auspice/issues/687).

This commit closes Auspice issue https://github.com/nextstrain/auspice/issues/696.