Update eslintrc and dependencies
I did these upgrades for two reasons:
1. We had GitHub security alerts regarding 'marked' to resolve
2. I was getting eslint errors in Atom for files within nextstrain.org/ and within nextstrain.org/static-site/. Looking at these it was clear that each had a different .eslintrc file and each specified different versions of eslint and eslint plugins and configs.

Here, I swapped in from auspice both the package.json requirements and the .eslintrc files. This fixed the eslint runtime errors and additionally I think it should be an improvement to be working with a consistent linter across auspice and these other bits of JS code.