add community redirect to server logic
Currently the redirect mechanisms aren't very streamlined. The server (server.js) must decide based on the route whether to serve the static site or auspice. Once the static site is served, it has its own redirect logic (`static-site/redirects.json`), but to get here the server must know to serve the static site. This should be improved in the future.