temporary fix for 404 page crashes
Currently the theme isn't being passed correctly to the 404 page, resulting in (e.g.) http://localhost:5000/about/blah crashing. This can only be replicated using the built version, not the development version of the static site. This commit fixes the issue by not using the theme in the <NavBar>, however a better solution would be to correctly supply the theme.