Implement server redirects in auspice
Previously, when an auspice client made a request to the
`nextstrain.org` server, we were injecting a `_url` key into the data to
send back to the client. The auspice client would then detect this magic
key and update the browser window path accordingly.

We wanted to move away from editing JSON files sent to the client. If
these changes are incorporated, the `nextstrain.org` server will check
the client request URL versus the URL of the best-guess dataset. If
these URLs do not match, then the server sends the client a redirect
(status code 301) to the full-length, updated URL.