Authentication for users using Amazon Cognito
This adds bare bones login/logout using Cognito's OAuth2 endpoints and
it's stock UI.  An advantage here is that all user management can be
done in the AWS Console (we don't have to build this out) and things
like password reset emails are handled for us.

Although we're not using it now, this positions us well for the future
to get temporary AWS credentials for our Cognito users to directly
access AWS resources.  As we plan to move dataset fetching into the
client in the future, this will be a requirement.

The UI/UX could be a lot better, but this seems good enough for an MVP.

For now, Heroku's session affinity plugin must be enabled.