charon: Consolidate the special handling of request URLs vis-à-vis sources
This removes special-casing of "live" vs. other sources from many places
and centralizes them into splitPrefixIntoParts() and
joinPartsIntoPrefix().

There is further refactoring that could be done here to improve
encapsulation, but this is a good first step that makes sources and
other code agnostic to the way sources are specified in the URLs.