charon/getDataset: Send 204 No Content responses for all source roots
This applies the change James made in a previous commit more
broadly.

A custom exception is used for future maintainability (for example,
detecting undefined classes is part of standard static analysis while
magic strings are not).  The exception is thrown at Dataset construction
time so it happens as early as possible.