install auspice globally
This commit returns to the behavior which existed before PR #27.

Eventually it would be preferable to have auspice as a (normal) dependency, but this is currently problematic. While it _mostly_ worked, it wasn't possible to install auspice from source using `npm install <local path>` and have the client-customisations work. This is important for testing purposes. 

See also: the previous commit.