add narratives section to splash page
includes changes to readme to indicate how to add new ones.
This is timed to coincide with the WNV paper which has a focus
on narrative storytelling. Eventually the number of community
and narrative cards will be too many for the splash page, but
we can tackle that situation when it arrives