eslint: Allow functions to be used before they're defined
Functions are hoisted so it is completely safe to use them before
they're defined.

Removing this ESLint restriction allows the very nice pattern of putting
higher-level functions at the top of a file and the lower-level
functions it calls at the bottom.  By putting callers before callees,
source files are well organized for top-to-bottom readability.