implement custom auspice splash page
This commit sets up a prototype implementation of a custom auspice splash
page, by adding an auspice built-time extension for the splash page.
No changes to the auspice source code were needed, this is all done as
a build-time extension, which I think is really cool.

The auspice splash page is accessed when the nextstrain.org router decides
a path should be handled by auspice, but auspice's request for the dataset
returns 500 or 204 (i.e. it doesn't exist). This commonly occurs in the
following cases (with this response code):
* nextstrain.org/doesnotexist (500)
* nextstrain.org/staging (204)
* nextstrain.org/<group_name> (204)
* nextstrain.org/community/<org_name> (500)
* nextstrain.org/community/<org_name>/<repo_name> (500 if
there is not an appropriate dataset in the repo)

Here I implement a custom nextstrain.org auspice splash page which makes
a API call to `/charon/getSourceInfo` (and implement the relevant handlers).
The response of this API call defines fields to be rendered, such as title,
byline etc (the exact structure of this is to be finalised, but note that
we can iterate on this after it is live but before publicising).

The intent of this (not yet implemented) is to allow different sources (e.g.
community GitHub repo, "nextstrain group") to define a JSON specifying this
which will be displayed via this splash page.