charon: Add general PublicGroupSource and PrivateGroupSource classes
Moves the shared bucket() definitions and visibleToUser() authz check,
leaving group sources able to be defined with just a name.

This also ensures that going forward source name and Cognito group name
match by default.  Matching is better in the long run than letting them
be different.  Note that this changes what group is used to authorize
access to /inrb-drc.  I've staged a new "inrb-drc" group in Cognito so
that this change can be deployed at any time without further
coordination.  Once it is deployed, we can delete the "inrb" Cognito
group.

It also ensures that source names and bucket names match, but INRB's
bucket is overridden to remain the same so that no one has to change
where they're putting files.