server: Serve an actual .ico instead of a .png for /favicon.ico
The "favicon" middleware serves a text/x-icon content-type, and browsers
expect the reported content-type to match the actual content.  I believe
the .ico will only be requested in very rare cases, but it's easy to fix
it now and forget about it.

For most pages, the favicon is specified by <link rel="shortcut icon">.
In Gatsby, this is an embedded data: URL; for Auspice, /favicon.png.

I used ImageMagick to `convert favicon.{png,ico}`.