server: Invert routing order so Gatsby is the default app, not Auspice
This lets us route dataset and narrative paths to Auspice (for core,
staging, community, and groups), but handle cases like group index pages
either in Gatsby or with a separate React app.

Most of this change is laying a foundation to make future changes
easier.  The only intended user-visible change is that Gatsby's 404 page
is now shown instead of Auspice's.  I think this is probably a slight
improvement—the splash page is a nicer error page than a dataset
index—but there's definitely still room to improve the 404 UX in the
future.

Resolves #49.