Fix visibleGroups is undefined error
In further testing of #74, we found an error when a user is logged-out
that prevented the splash page of nextstrain.org from loading.

Move the logic for determining which groups are visible to a user into
the `/whoami` endpoint. Send the resulting `visibleGroups` as an
additional key in the json load sent by this endpoint. Handle the
additional `visibleGroups` object accordingly.