Make public & private nextstrain groups available to all users
This modifies the code such that the `visibleGroups` prop injected by `<UserDataWrapper>` contains (a) all public groups and (b) all private groups for which the user has the correct credentials (in this case, a matching cognito group). The entries of the `visibleGroups` array are now objects with `name` and `private` properties.

The /users/<userName> page now only shows those groups which are private, which matches the text on that page.