Identify private group tiles with padlock icon
The `<UserGroups>`component shows both private & public (nextstrain) groups. This adds a padlock icon onto private groups & changes the title & subtext of the component to better reflect what's displayed.
Note that this component is only displayed _if_ a user is logged in, however it can easily be modified to display public groups for non-logged in users.