disable `gatsby-plugin-catch-links`
This was catching relative links which should got to auspice (e.g. the zika tile on the splash page) but modifying the link to go to a gatsby 404 (the server never saw the get request and so couldn't send to auspice). This plugin is generally problematic for us & needs attention.