Move users page to sections
We want to create the users "page" via a `createPage` call so that we can use it /users/:username URLs. When this was located within the `/pages` directory gatsby automatically created a page at `/users` which loaded fine, but the `createPage` functionality was lost (i.e. /users/userName wouldn't load). By moving this file gatsby no longer automatically creates a URL for it and we get back our desired behaviour via `createPage`.