Limit FileStore retries. Closes #39
When no REDIS_URL env var is set then we're running on localhost (or perhaps in a heroku review-app) and we store the session info in a FileStore. This was prone to spurious, but harmless, warnings printed in the terminal which was problematic for development. We no longer retry to access the session on disk (if it fails) which prevents these warnings.