change auspice folder structure to prevent auspice build bugs
This moves the "auspice" server code to `./src` so that all the auspice client code is in it's own directory (./auspice-client`). This helps prevent `auspice build` pulling in code it shouldn't and which it doesn't correctly transpile. Note that this is a work-around for the deeper problem explained at https://github.com/nextstrain/auspice/issues/902.

Moving server code to `./src` makes sense now that the nextstrain.org capabilities have advanced. Some file renaming may be in order here for clarity, and we should consider moving `authn.js` and `redirects.js` into `./src`.