S3Source: return `logo.png` as avatar if found in S3 bucket
This will allow Nextstrain groups to customize their logo/avatar on
the groups splash page by uploading `logo.png` to their S3 bucket.

The image is fetched server-side and converted to base64 to be displayed
client-side via data URL. This is done to allow the display of the
logo for private groups that use private S3 buckets.