Add external links to docs pages in sidebar
We've wanted the ability to provide external links to docs pages -- for instance the new nCoV tutorial which is hosted by github pages for the ncov repo.

This commit sets up the ability to do this for a curated set of entries (in a YAML file), and indicates that they will go to an external doman via a icon next to them in the sidebar.