Move to using `npm ci`
We use npm to install project dependencies in two places: the root level and within the `static-site` directory (via `build.sh` or `npm run build`). Moving to `npm ci` allows (a) developers to all be working from the same deterministic set of dependencies and (b) the automatic deploys (heroku, travis) to use the same dependencies as us.