Use node 13.6
This updates travisCI to use the version specified in package.json, which heroku will run. This results in a CI which better matches our actual deploy environment.

Also changed the `engines` declaration of the static-site to match the root level declaration (I don't believe anything was / is using the static-site engines declaration, but it should match the root level as we are running both from the same environment).