Send error messages in the response body, not the status line
The HTTP status line message is not intended for arbitrary messages and
not suitable for freeform text due to character restrictions.

Using statusMessage this way has caused nextstrain.org server crashes
when the field is assigned an error containing a newline, as newlines
are not allowed.

Resolves #201.

Related to <https://github.com/nextstrain/auspice/issues/1180>.