Allow API requests to be handled by the main server in dev mode
Certain parts of the gatsby site depend on API calls to handlers outside of gatsby (they're implemented by the nextstrain.org server). This is a development hand-brake as it requires building the gatsby site and running the nextstrain.org server to test functionality.

Here I introduce a simple solution where, when running the gatsby development server (on localhost:8000), the API request is made to localhost:5000, allowing the developer to run the main server in a separate process. This required enabling CORS for the specific request from the specific origin which the gatsby  dev server runs from. 

This provides a general example of how we can address issue #158.