Allow community builds to be stored using git-lfs
The API we were using to access the datasets behind nextstrain community didn't work if the file was stored using git-lfs (the response was a 200 with the contents of the git-lfs pointer). We now switch to a different GitHub API call which allows us to access both "normal" github files and those stored with git-lfs.

Underneath, this new API returns a 302 redirect to one of two different APIs. If the file is "normal", the redirect is to the previousl API we were using (raw.githubusercontent.com), if it's git-lfs then it redirects the request to the media.githubusercontent.com API.