Rebuild package-lock
This was motivated by a change in the lock-file introduced in 04b7da8
which, for reasons unknown to me, changed the resolved path (.tgz) of
the `eslint-visitor-keys` package (a dependency of eslint, among others)
to a completely different package (`camelize`). Rebuilding the lock-
file restored expected behavior.

This change in package caused unexpected and hard-to-interpret eslint
errors, which this commit now fixes.

See https://github.com/nextstrain/nextstrain.org/pull/225#issuecomment-704677191
for more.