Refactor secondTreeOptions
Refactor the computation of second tree options data by adding it as a
method of the Source classes.

Note that it still uses globally computed data for now. While stopping
or centralizing the usage of global variables is a desirable direction
to take this codebase, this commit is just moving the functionality to a
place it's more expected.