Return 4XX errors on bad client requests
An invalid query (i.e. where getDataset is missing a prefix) is a client
request error, not an internal server error. Instead of a 500 error,
return a 400 error when the client sends a request for a URL/prefix that
cannot be parsed. In future commits, we'll also return a 404 error when
the requested data don't exist, following this same pattern.