first draft of cron job to update sars-cov-2 search;
this is a Github action which runs
scripts/collect-search-results.js
every weekday as close as possible
to the time by which we have usually
published new sars-cov-2 (ncov) builds
to s3. More jobs in this workflow or
separate ones with separate timelines
can be added for other pathogens we
want to maintain search pages for.
If we want to make this 'smarter' we
can use something like aws lambdas
which could automatically run upon
updating of certain files in an s3
bucket