install deps for workflow in same step as running script;
you need to explicitly keep
changes to the environment
from one step to the next
within a github action job
by using the PATH variable
if I remember correctly.
Avoiding dealing with this
for now by sticking it all
in the same step, since if
we want to update other
search databases in addition
to sars-cov-2, they'll either
be on the same timeline and
can be tacked onto this step,
or a separate timeline and
will therefore need a separate
workflow. Unless we want some
updates to fail within a workflow
and not have it prevent the rest
(which seems smart).