charon: Remove indirection of fetchUrls object
It is simpler for the getDataset handler to use the Dataset object
directly and pass it around when needed.

parsePrefix() is now a lot leaner with a focused purpose: given a
prefix, return the relevant Source and Dataset objects, along with the
canonicalized version of the prefix.  Notably, request details like the
query params are no longer needed by it.