charon: Redirect to the fully-resolved prefix more robustly
Compare (canonicalized) prefixes, not URLs, because that's more
precisely what we're concerned about here.

Construct the relative URL for redirection using proper URL handling.
The previous method for constructing the new URL was problematic for
several reasons, including that whether other query params were passed
along or not with the baseUrl depended on where they were relative to
the prefix param.  Also, while unlikely in this case, if a prefix value
appeared elsewhere in the URL earlier than the query param, it all broke
because of the split().