[build-map] cluster community builds on the map;
This uses a react-mapbox-gl component to cluster
builds on the map. Only community builds are
clustered in this way. Clicking a cluster will
zoom to the bounds of the builds it contains.
Cluster tooltips don't show the contents of the
cluster (this is possible, just adds complexity
to implementation). Important to note that
clusters will never separate if two markers
(builds) have the same coordinates. Since
programmatically adding small differences to
identical coordinates is non-trivial, we do
this manually in the yaml file for now. We
should add a console warning so that this
doesn't happen by mistake.