[build-map] increase map marker jitter;
this is done manually by just adjusting
coordinates for now. this is pretty easy
to do if you are willing to spend the time
fiddling with it. writing a programmatic
way to add jitter would be better since
it doesn't require folks who are adding
builds to consider this at all, but it might
be tricky to make sure there were no
collisions. Even better would be allowing
duplicate coordinates and using cluster
tooltips to allow you to see what's "inside"
each cluster and navigate to those builds
through the tooltip without having to separate
them on the map. Doing that (within the current
implementation) requires managing react
component state of the cluster, so we are
skipping it for now.