generalize builds component;
in anticipation of using it for any listing of
builds, not specific to sars-cov-2. This
picks up where that left off in
anticipation of using it for a flu builds
page. Specifically this commit allows
the build hierarchy to be defined by custom
field names in a build catalogue yaml (e.g.
static-site/content/allSARS-CoV-2Builds.yaml)
. These field names are passed to the builds
component so it can create the hierarchy
according to them.