create /influenza page;
This creates a page at /influenza
analogous to /sars-cov-2 but for flu.
Note that this page isn't called /flu
because of conflict with existing redirects.
In addition to creating a useful hub for
gathering resources about flu, this can
serve as an example of how it should
hopefully now be easier to create more
pages like this that show off all that
Nextstrain has to offer on a particular
pathogen or topic spanning pathogens
(e.g.  a /groups page that lists all
nextstrain groups and links to docs
about groups etc). In many cases
these pages can serve as a way to
dedicate more UI real estate to some
things or sections we show off on the
splash page and offer a more customized
landing page for a particular topic.