Improve layout of dataset selection
This uses react-styled-flexboxgrid for a responsive grid layout. We use Bootstrap elsewhere in nextstrain.org to accomplish a similar outcome, however, Bootstrap css and className="col-md-1" don't interact well with the use of styled-components. Here, I've opted for a Flexbox solution. This adds a dependency to package.json, but allows for an easy, expressive grid and I think should be generally useful. I compared a few similar options for this and this one was my favorite.