filterable dataset list on /sars-cov-2
instead of dropdowns, use the filterable
dataset-select component. Because the
urls for these datasets are less uniform
than e.g. the nextstrain-maintained /influenza
listing, the keywords give you groupings
that are often just one dataset.
This is not too bad, except in some cases
where use shortened urls that point
to a splash page or redirect to the real
dataset. These urls dont give useful
keywords so should be changed
(something we should do regardless
to make urls in the YAML more uniform).

This makes the date updated column
optional through a prop passed to hide it.
This is as opposed to conditionally
showing the value in the column
on a row-by-row basis. I went with the
simpler approach since we don't expect
any datasets to have this property in this
case.