[gatsby] Update how influenza pages are configured
Here we switch back to useing `matchPath` rather than
gatsby-plugin-create-client-paths to keep things consistent.
For unknown reasons, if the component is named `influenza.jsx` we lose
the matchPath functionality. Therefore in a future commit we should
simultaneously change the path & matchPath to "flu", change the page
back to `influenza.jsx`, and update the routing logic of the server
so that URLs starting with /flu which define valid (core) datasets,
and only these, are sent to auspice, with the rest falling through to
Gatsby.