Only filter to valid values
The filtering logic is lifted from Auspice, where we can
guarantee that applied filters are valid (i.e. they exist as options).
Here we don't have this guarantee, as we can define keyword filters
via the URL. This commit filters out values which don't exist as
options.

The URL is left unmodified, however ideally this would be updated.