authn/bearer: Remove support for "scope" option
We have no need for this option, which exists only to add a static,
"scope=…" field to the WWW-Authenticate header sent with 401
Unauthorized responses generated by this strategy.  Such responses are
only sent for invalid tokens since no "Authorization: Bearer …" header
in the request means we fallback to an anonymous user instead of
immediately rejecting.  Our app-site error handler will generate a 401
if necessary if our authz code rejects the anonymous user later in the
request processing.

Additionally, the documentation for this option is potentially
misleading as it can imply (as I first read it) that this strategy will
actually validate that the token scopes match.  It doesn't and indeed
can't as the tokens are opaque to this code, but things that are
misleading don't pair well with security code.

Similarly, supporting more authn code paths than we need only makes it
harder to reason about the security of the system as a whole.

Noticed while reading through this for details of the WWW-Authenticate
response header this code generates.