authn/bearer: Add a passIfMissing option for when no Authorization header is present
Passing is how Passport supports moving to the next request
handler/middleware with next().

Chaining with the passport-anonymous strategy is the typically-suggested
way to do something similar, but it has the downside of both hiding all
authentication errors from the client and not distinguishing between
invalid credentials and missing credentials.