Update URLs to reflect parent page
This moves us towards a system whereby we define a set of pagenames
(static or dynamic) and any extra data in the path is removed upon
loading.

For instance, given the "influenza" page (as used in this commit), then
requests to /influenza/a/b/c will load the influenza page and have their
URL changed to /influenza. The fields a, b & c are interpreted as
potential filter queries of the dataset listing on that page.

Note that this only applies for pages that our router passes to Gatsby,
i.e. those which are not sent to Auspice.