authn: Proper error handling in our user verification routine
Passport's API uses callback-passing style and so our async function
must explicitly catch errors in order to call Passport's done() callback
appropriately.  Previously, errors during authn caused an unexpected
server error.

Changes the redirection target for failures to avoid a redirect loop.
We still don't have any framework in place for displaying dynamic
messages to users (such as with connect-flash and server-rendered
pages), so keep punting on that for now.  It's bad UX, but hopefully
will be pretty rarely encountered.  We can improve it when the necessary
framework is in place.