authn: Accept Bearer tokens in Authorization headers
Prepares the path for the Nextstrain CLI to make authenticated requests
to existing (and future) API endpoints.  Only tokens minted for the
Nextstrain CLI are accepted for now, but we can easily add additional
clients (first- or third-party) later.

If an Authorization header is present, then it must container a valid id
token issued from our Cognito user pool.  Errors are returned if the
token is invalid.  No session will be persisted for requests with an
Authorization header.

If an Authorization header is not present (or is completely empty), then
request handling passes on to the previous behaviour, which restores any
authenticated user session that's present.