Dataset visualization as prop to <DatasetSelect>
This exposes the ability for pages (e.g. sars-cov-2, influenza)
to define which elements are rendered by <DatasetSelect>, and
inject custom elements.

For instance, the geographic map of SARS-CoV-2 datasets is now
handed to <DatasetSelect> as a prop for rendering.

This commit was based off an original design by eharkins

Co-authored-by: eharkins <eli.harkins@gmail.com>