Rendered columns are also filters
Columns are currently supplied to the <DatasetSelect> for rendering
in a table. Here we also expose these as filtering options.
The first column (the dataset) is special cased and broken
into keywords, thus allowing (e.g.) us to match "h3n2" for a dataset
flu/seasonal/h3n2/ha/2y.

This commit rebases previous work by @harkins onto a refactored
version of the <DatsetSelect>, and extends it to require datasets to
match the union of filterValues for each filterType specified.

Co-authored-by: eharkins <eli.harkins@gmail.com>