gatsby create page fix;
this fixes unintended rendering of certain pages in the dir:
static-site/src/pages. Gatsby auto-creates pages
from js,jsx pages in this directory. This issue was raised here:
https://github.com/nextstrain/nextstrain.org/pull/299#issuecomment-820696299
In some cases, we want to call createPage in static-site/gatsby-node.js
to create multiple other pages under a certain URL, so this
puts those in static-site/src/sections as a workaround to avoid
rendering of pages at a URL matching the exact name of the
file (such as nextstrain.org/influenza-page). These
files are intended to help render pages matching
a specified URL (such as nextstrain.org/influenza/*).