Refactor server routing to handle missing datasets
Previously, URLs which look like datasets were sent to Auspice,
which would attempt to fetch the dataset file &, if this fails,
render a splash page. This included routes like `/groups/blab`

This commit introduces the framework to verify if such routes have
a corresponding dataset, and use this to route requests to Gatsby
or Auspice. Ths is done via express middleware. Furthermore, paths
for which there are no datasets can be sent a specific Gatsby
page rather than the (default) 404.

This commit only modifies routes which source datasets from our
"core" source. Generally, if the dataset is missing we return the
gatsby 404 page and don't modify the URL. If the requested (but non
existing) dataset starts with "flu", "ncov" or "sars-cov-2" we instead
render a "closer" page: /influenza or /sars-cov-2. In this case the
URL is updated (by Gatsby, client side), however this should not be
the case.